backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
psycopg2-binary
pydantic>=2
PyJWT[crypto]
python-dotenv
supabase
# RAG Pipeline Dependencies
sentence-transformers
langchain
langchain-community
langchain-openai
pypdf
python-multipart
numpy
faiss-cpu
tiktoken
# pgvector support
psycopg2-binary
